log likelihood orthogonal

log likelihood orthogonal

loglik_math(
  param_vector,
  env_dat,
  occ,
  opt = NULL,
  num_threads = RcppParallel::defaultNumThreads(),
  negative = TRUE
)

Arguments

param_vector

A vector with parameters to test in a log-likelihood function of the species distribution model.

env_dat

An array with environmental data, includes the time series in each location for each environmental variable

occ

A presence absence vector

opt

An optional vector to mute a specific parameter to evaluate. It is further used to profile the optimization

num_threads

The number of threads to run the main function

negative

Logical. If true returns the negative of the log-likelihood function. Flag should be set when optimize.

Value

A value with the log of the likelihood for a given set of parameters, environmental variables and presence-absences occurrences.

Examples

env_dat <- envdat_ex
occ <- occExample
params <- param_table_example[5, ]
param_vector <- params
loglik_math(params, env_dat, occ)
#> [1] 2596.381

opt_a <- c(
  mu1 = NA_real_,
  mu2 = 0,
  sigl1 = NA_real_,
  sigl2 = NA_real_,
  sigr1 = NA_real_,
  sigr2 = NA_real_,
  ctil = NA_real_,
  pd = NA_real_,
  opar_1 = NA_real_
)
loglik_math(params, env_dat, occ, opt = opt_a)
#> [1] 16236.4