startparms Starting parameters for the optimization. Given a set of ranges of values of environmental variables where the presence of a species occurs,generates a Latin hypercube design for the parameters based on the Sobol' low-discrepancy sequence.

startparms Starting parameters for the optimization. Given a set of ranges of values of environmental variables where the presence of a species occurs,generates a Latin hypercube design for the parameters based on the Sobol' low-discrepancy sequence.

start_parms(env_dat, quant_vec = c(0.1, 0.5, 0.9), num_starts = 100)

Arguments

env_dat

The environmental array given the observed occurrences

quant_vec

A vector of quantiles. Are the bounds of the range to create the hypercube

num_starts

The number of samples of the hypercube

Value

A data frame with samples for each parameter to optimize

Examples

envdat_ex_occ <- envdat_ex[, , occExample == 1]
start_parms(envdat_ex_occ)
#> # A tibble: 100 × 9
#>      mu1   mu2   sigl1    sigl2    sigr1    sigr2   ctil     pd  o_par1
#>    <dbl> <dbl>   <dbl>    <dbl>    <dbl>    <dbl>  <dbl>  <dbl>   <dbl>
#>  1  9.59  4.00  0.361   0.356    0.171    0.616   -1.26   0.858 -1.62  
#>  2 11.8   5.29  1.05   -0.338    0.864   -0.0773  -0.624 -1.34   1.52  
#>  3 12.9   3.36  0.708   0.00895  1.21     0.269   -0.944  1.96  -0.0491
#>  4 10.7   4.65  0.0145 -0.684    0.518   -0.424   -1.58  -0.240  3.09  
#>  5 11.2   3.04  0.881   0.529    1.04    -0.251   -1.42  -1.89  -2.41  
#>  6 13.4   4.32  0.188  -0.164    0.344    0.443   -0.784  0.309  0.736 
#>  7 12.3   3.68 -0.159   0.182   -0.00210  0.0960  -1.10  -0.790 -0.834 
#>  8 10.1   4.97  0.534  -0.511    0.691    0.789   -1.74   1.41   2.31  
#>  9 10.4   3.52 -0.0721 -0.0777   0.778    0.702   -1.18  -1.06  -2.80  
#> 10 12.6   4.81  0.621   0.615    0.0845   0.00934 -0.544  1.13   0.344 
#> # ℹ 90 more rows